Fix Menu receiving no longer valid className prop, improve stories
