--- stories/components/styles/menu_story.less
@@ -1,41 +1,0 @@
-.story {
-    font-family: Arial;
-
-    button[aria-haspopup='menu'] svg {
-        fill: currentcolor;
-        width: 0.65em;
-        height: 0.65em;
-        margin-left: 0.5em;
-    }
-    ul {
-        list-style: none;
-        padding: 0;
-    }
-    .item_box {
-        min-width: 300px;
-        display: flex;
-        border: 1px solid grey;
-        padding: 4px 10px;
-        margin: 4px 0;
-        user-select: none;
-        &:hover {
-            background-color: #efefef;
-        }
-        &.menu_open {
-            background-color: #cfcfcf;
-        }
-        .item_name {
-            flex: 1;
-            display: flex;
-            align-items: center;
-        }
-    }
-
-    .menu_item_icon {
-        padding-right: 10px;
-    }
-
-    .menu_item_label {
-        flex: 1;
-    }
-}
